♻️ [Refac] 플로팅버튼 구조 개선
position으로 위치하도록 설정
